requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
Flask==3.0.2